Bullet Drop
Box and Player has a projectile.
Create a function in the third person character.
that accepts a start position, a velocity, and a MaxSimTime
and returns whether we hit something, and if so, at what time (what step?), and where?
if hit was successful, a sound plays and effect appears (at the correct time and location).